• Restyle the bottom actions. • Show press delete key to delete the object message. • Prevent deletion in simulation mode. • Fix rack items scrolling.

+++ js/view_models/gate_panel.js
@@ -11,6 +11,15 @@ let gatePanelViewModel = function (runningMode, shouter, state, gfxEventHandler,
     self.activeGateCol = -1;
 
     self.add = function (row, col) {
+        if (runningMode() !== RUNNING_MODE.DESIGN) {
+            shouter.notifySubscribers({
+                text: "This action is allowed in design mode only!",
+                type: MSG_TYPE.ERROR
+            }, SHOUT.MSG);
+
+            return false;
+        }
+
         let isFacilityFree = state.map.isFacilityFree(row, col);
 
         if (self.editing()) {
@@ -104,6 +113,15 @@ let gatePanelViewModel = function (runningMode, shouter, state, gfxEventHandler,
     };
 
     self.delete = function (row, col) {
+        if (runningMode() !== RUNNING_MODE.DESIGN) {
+            shouter.notifySubscribers({
+                text: "This action is allowed in design mode only!",
+                type: MSG_TYPE.ERROR
+            }, SHOUT.MSG);
+
+            return false;
+        }
+
         let fac = state.map.getSpecificFacility(row, col, MAP_CELL.GATE);
 
         state.map.deleteObject(row, col, fac);
