+++ js/view_models/rack_panel.js
@@ -19,6 +19,15 @@ let rackPanelViewModel = function (runningMode, shouter, state, gfxEventHandler,
     self.activeRackCol = -1;
 
     self.add = function (row, col) {
+        if (runningMode() !== RUNNING_MODE.DESIGN) {
+            shouter.notifySubscribers({
+                text: "This action is allowed in design mode only!",
+                type: MSG_TYPE.ERROR
+            }, SHOUT.MSG);
+
+            return false;
+        }
+
         let isFacilityFree = state.map.isFacilityFree(row, col);
 
         if (self.editing()) {
@@ -118,6 +127,15 @@ let rackPanelViewModel = function (runningMode, shouter, state, gfxEventHandler,
     };
 
     self.delete = function (row, col) {
+        if (runningMode() !== RUNNING_MODE.DESIGN) {
+            shouter.notifySubscribers({
+                text: "This action is allowed in design mode only!",
+                type: MSG_TYPE.ERROR
+            }, SHOUT.MSG);
+
+            return false;
+        }
+
         let fac = state.map.getSpecificFacility(row, col, MAP_CELL.RACK);
 
         state.map.deleteObject(row, col, fac);
